+++ data/potato.go
@@ -73,7 +73,7 @@ func (p *Potato) WriteStage1(c chan string, target, traits string) {
 	<drop event="Destroy" name="plantedPotato1_%s" count="1"/>
 	<property name="CreativeMode" value="Player"/>
 	<property name="CustomIcon" value="plantedPotato1"/>
-	<property name="DescriptionKey" value="plantedPotato1_%s"/>
+	<property name="DescriptionKey" value="plantedPotato1_%sDesc"/>
 	<property name="Extends" value="cropsGrowingMaster" param1="CustomIcon"/>
 	<property name="Group" value="%s"/>
 	<property name="Model" value="Entities/Plants/potato_plant_sproutPrefab"/>
@@ -104,7 +104,7 @@ func (p *Potato) WriteStage3(c chan string, traits string) {
 	<property name="CreativeMode" value="Dev"/>
 	<property name="CustomIcon" value="plantedPotato3Harvest"/>
 	<property name="CustomIconTint" value="ff8000"/>
-	<property name="DescriptionKey" value="plantedPotato3_%s"/>
+	<property name="DescriptionKey" value="plantedPotato3_%sDesc"/>
 	<property name="DisplayInfo" value="Description"/> <!-- also valid: "Name" -->
 	<property name="DisplayType" value="blockMulti"/>
 	<property name="Extends" value="cropsHarvestableMaster"/>